+++ src/amqp-base-client.ts
@@ -4,7 +4,7 @@ import { AMQPMessage } from './amqp-message.js'
 import { AMQPView } from './amqp-view.js'
 import { Logger } from './types'
 
-const VERSION = '2.1.1'
+const VERSION = '3.0.0'
 
 /**
  * Base class for AMQPClients.
@@ -19,6 +19,7 @@ export abstract class AMQPBaseClient {
   channels: AMQPChannel[]
   protected connectPromise?: [(conn: AMQPBaseClient) => void, (err: Error) => void]
   protected closePromise?: [(value?: void) => void, (err: Error) => void]
+  protected onUpdateSecretOk?: (value?: void) => void
   closed = true
   blocked?: string
   channelMax = 0
@@ -72,21 +73,7 @@ export abstract class AMQPBaseClient {
 
     const channel = new AMQPChannel(this, id)
     this.channels[id] = channel
-
-    let j = 0
-    const channelOpen = new AMQPView(new ArrayBuffer(13))
-    channelOpen.setUint8(j, 1); j += 1 // type: method
-    channelOpen.setUint16(j, id); j += 2 // channel id
-    channelOpen.setUint32(j, 5); j += 4 // frameSize
-    channelOpen.setUint16(j, 20); j += 2 // class: channel
-    channelOpen.setUint16(j, 10); j += 2 // method: open
-    channelOpen.setUint8(j, 0); j += 1 // reserved1
-    channelOpen.setUint8(j, 206); j += 1 // frame end byte
-    return new Promise((resolve, reject) => {
-      this.send(new Uint8Array(channelOpen.buffer, 0, 13))
-        .then(() => channel.promises.push([resolve, reject]))
-        .catch(reject)
-    })
+    return channel.open()
   }
 
   /**
@@ -112,6 +99,25 @@ export abstract class AMQPBaseClient {
     return new Promise((resolve, reject) => {
       this.send(new Uint8Array(frame.buffer, 0, j))
         .then(() => this.closePromise = [resolve, reject])
+        .catch(reject)
+    })
+  }
+
+  updateSecret(newSecret : string, reason : string) {
+    let j = 0
+    const frame = new AMQPView(new ArrayBuffer(4096))
+    frame.setUint8(j, 1); j += 1 // type: method
+    frame.setUint16(j, 0); j += 2 // channel: 0
+    frame.setUint32(j, 0); j += 4 // frameSize
+    frame.setUint16(j, 10); j += 2 // class: connection
+    frame.setUint16(j, 70); j += 2 // method: update-secret
+    j += frame.setLongString(j, newSecret) // new secret
+    j += frame.setShortString(j, reason) // reason
+    frame.setUint8(j, 206); j += 1 // frame end byte
+    frame.setUint32(3, j - 8) // update frameSize
+    return new Promise((resolve, reject) => {
+      this.send(new Uint8Array(frame.buffer, 0, j))
+        .then(() => this.onUpdateSecretOk = resolve)
         .catch(reject)
     })
   }
@@ -282,6 +288,7 @@ export abstract class AMQPBaseClient {
                     .catch(err => this.logger?.warn("Error while sending Connection#CloseOk", err))
                   this.onerror(err)
                   this.rejectConnect(err)
+                  this.onUpdateSecretOk?.()
                   break
                 }
                 case 51: { // closeOk
@@ -307,6 +314,12 @@ export abstract class AMQPBaseClient {
                   delete this.blocked
                   break
                 }
+                case 71: { // update-secret-ok
+                  console.info("AMQP connection update secret ok")
+                  this.onUpdateSecretOk?.()
+                  delete this.onUpdateSecretOk
+                  break
+                }
                 default:
                   i += frameSize - 4
                   this.logger?.error("unsupported class/method id", classId, methodId)
@@ -317,12 +330,12 @@ export abstract class AMQPBaseClient {
               switch (methodId) {
                 case 11: { // openok
                   i += 4 // reserved1 (long string)
-                  channel.resolvePromise(channel)
+                  channel.resolveRPC(channel)
                   break
                 }
                 case 21: { // flowOk
                   const active = view.getUint8(i) !== 0; i += 1
-                  channel.resolvePromise(active)
+                  channel.resolveRPC(active)
                   break
                 }
                 case 40: { // close
@@ -351,7 +364,7 @@ export abstract class AMQPBaseClient {
                 case 41: { // closeOk
                   channel.setClosed()
                   delete this.channels[channelId]
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 default:
@@ -366,7 +379,7 @@ export abstract class AMQPBaseClient {
                 case 21: // deleteOk
                 case 31: // bindOk
                 case 51: { // unbindOk
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 default:
@@ -381,25 +394,25 @@ export abstract class AMQPBaseClient {
                   const [name, strLen] = view.getShortString(i); i += strLen
                   const messageCount = view.getUint32(i); i += 4
                   const consumerCount = view.getUint32(i); i += 4
-                  channel.resolvePromise({ name, messageCount, consumerCount })
+                  channel.resolveRPC({ name, messageCount, consumerCount })
                   break
                 }
                 case 21: { // bindOk
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 case 31: { // purgeOk
                   const messageCount = view.getUint32(i); i += 4
-                  channel.resolvePromise({ messageCount })
+                  channel.resolveRPC({ messageCount })
                   break
                 }
                 case 41: { // deleteOk
                   const messageCount = view.getUint32(i); i += 4
-                  channel.resolvePromise({ messageCount })
+                  channel.resolveRPC({ messageCount })
                   break
                 }
                 case 51: { // unbindOk
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 default:
@@ -411,12 +424,12 @@ export abstract class AMQPBaseClient {
             case 60: { // basic
               switch (methodId) {
                 case 11: { // qosOk
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 case 21: { // consumeOk
                   const [consumerTag, len] = view.getShortString(i); i += len
-                  channel.resolvePromise(consumerTag)
+                  channel.resolveRPC(consumerTag)
                   break
                 }
                 case 30: { // cancel
@@ -444,7 +457,7 @@ export abstract class AMQPBaseClient {
                 }
                 case 31: { // cancelOk
                   const [consumerTag, len] = view.getShortString(i); i += len
-                  channel.resolvePromise(consumerTag)
+                  channel.resolveRPC(consumerTag)
                   break
                 }
                 case 50: { // return
@@ -492,7 +505,7 @@ export abstract class AMQPBaseClient {
                 }
                 case 72: { // getEmpty
                   const [ , len]= view.getShortString(i); i += len // reserved1
-                  channel.resolvePromise(null)
+                  channel.resolveRPC(null)
                   break
                 }
                 case 80: { // confirm ack
@@ -502,7 +515,7 @@ export abstract class AMQPBaseClient {
                   break
                 }
                 case 111: { // recoverOk
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 case 120: { // confirm nack
@@ -521,7 +534,7 @@ export abstract class AMQPBaseClient {
               switch (methodId) {
                 case 11: { // selectOk
                   channel.confirmId = 1
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 default:
@@ -535,7 +548,7 @@ export abstract class AMQPBaseClient {
                 case 11: // selectOk
                 case 21: // commitOk
                 case 31: { // rollbackOk
-                  channel.resolvePromise()
+                  channel.resolveRPC()
                   break
                 }
                 default: