+++ src/amqp-socket-client.ts
@@ -1,6 +1,6 @@
 import { AMQPBaseClient } from './amqp-base-client.js'
 import { AMQPError } from './amqp-error.js'
-import { AMQPTlsOptions } from './amqp-tls-options'
+import type { AMQPTlsOptions } from './amqp-tls-options.js'
 import { AMQPView } from './amqp-view.js'
 import { Buffer } from 'buffer'
 import * as net from 'net'
@@ -150,7 +150,10 @@ export class AMQPClient extends AMQPBaseClient {
 
   protected override closeSocket(): void {
     this.closed = true
-    if (this.socket) this.socket.end()
+    if (this.socket) {
+      this.socket.end()
+      this.socket.destroy()
+    }
     this.socket = undefined
   }
 }